用户基本信息逻辑
