+++ src/main/java/com/chinagreentown/dmp/Mapper/UsrBasAttrMapper.java
@@ -1,0 +1,20 @@
+package com.chinagreentown.dmp.Mapper;
+
+import com.chinagreentown.dmp.api.RowMapper;
+import com.chinagreentown.dmp.pojo.UsrBasAttrPojo.attr;
+import com.chinagreentown.dmp.pojo.UsrCNetBhvrPojo.bhvr;
+import com.chinagreentown.dmp.util.BeanUtil;
+import org.apache.hadoop.hbase.client.Result;
+
+/**
+ * Created by yun on 2017/4/18.
+ */
+public class UsrBasAttrMapper implements RowMapper<attr> {
+
+
+    @Override
+    public attr mapRow(Result result, int rowNum) throws Exception {
+        attr o = (attr) BeanUtil.mapRow(result, attr.class);
+        return o;
+    }
+}
