+++ src/main/java/com/chinagreentown/dmp/controller/DataCenterController.java
@@ -1,6 +1,5 @@
 package com.chinagreentown.dmp.controller;
 
-import com.chinagreentown.dmp.pojo.UsrBasAttrPojo;
 import com.chinagreentown.dmp.service.QueryService;
 import com.chinagreentown.dmp.util.FakeData;
 import com.google.common.collect.Maps;
