--- src/main/java/com/chinagreentown/dmp/pojo/UsrBasAttrPojo.java
@@ -1,217 +1,0 @@
-package com.chinagreentown.dmp.pojo;
-
-/**
- * Created by yun on 2017/4/17.
- *
- * @description 对应 usr_bas_attr（用户基础属性表）
- */
-public class UsrBasAttrPojo {
-
-    public class attr {
-        //名字
-        private String name;
-        //性别
-        private String gender;
-        //年龄
-        private String age;
-        //宽带账号
-        private String bd_account;
-        //名族
-        private String ethnic;
-        //手机号
-        private String tel;
-        //email
-        private String mail;
-        //微信
-        private String weixin;
-        //qq
-        private String qq;
-        //证件
-        private String papers;
-        //职业
-        private String profession;
-        //学历
-        private String education;
-        //职称
-        private String pro_title;
-        //政治面貌
-        private String pol_status;
-        //资产信息 房产
-        private String house_property;
-        //资产信息车
-        private String car_property;
-        //社交信息
-        private String social_info;
-        //用户信息
-        private String member_info;
-        //家庭结构
-        private String family_info;
-        //健康信息
-        private String health_info;
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public String getGender() {
-            return gender;
-        }
-
-        public void setGender(String gender) {
-            this.gender = gender;
-        }
-
-        public String getAge() {
-            return age;
-        }
-
-        public void setAge(String age) {
-            this.age = age;
-        }
-
-        public String getBd_account() {
-            return bd_account;
-        }
-
-        public void setBd_account(String bd_account) {
-            this.bd_account = bd_account;
-        }
-
-        public String getEthnic() {
-            return ethnic;
-        }
-
-        public void setEthnic(String ethnic) {
-            this.ethnic = ethnic;
-        }
-
-        public String getTel() {
-            return tel;
-        }
-
-        public void setTel(String tel) {
-            this.tel = tel;
-        }
-
-        public String getMail() {
-            return mail;
-        }
-
-        public void setMail(String mail) {
-            this.mail = mail;
-        }
-
-        public String getWeixin() {
-            return weixin;
-        }
-
-        public void setWeixin(String weixin) {
-            this.weixin = weixin;
-        }
-
-        public String getQq() {
-            return qq;
-        }
-
-        public void setQq(String qq) {
-            this.qq = qq;
-        }
-
-        public String getPapers() {
-            return papers;
-        }
-
-        public void setPapers(String papers) {
-            this.papers = papers;
-        }
-
-        public String getProfession() {
-            return profession;
-        }
-
-        public void setProfession(String profession) {
-            this.profession = profession;
-        }
-
-        public String getEducation() {
-            return education;
-        }
-
-        public void setEducation(String education) {
-            this.education = education;
-        }
-
-        public String getPro_title() {
-            return pro_title;
-        }
-
-        public void setPro_title(String pro_title) {
-            this.pro_title = pro_title;
-        }
-
-        public String getPol_status() {
-            return pol_status;
-        }
-
-        public void setPol_status(String pol_status) {
-            this.pol_status = pol_status;
-        }
-
-        public String getHouse_property() {
-            return house_property;
-        }
-
-        public void setHouse_property(String house_property) {
-            this.house_property = house_property;
-        }
-
-        public String getCar_property() {
-            return car_property;
-        }
-
-        public void setCar_property(String car_property) {
-            this.car_property = car_property;
-        }
-
-        public String getSocial_info() {
-            return social_info;
-        }
-
-        public void setSocial_info(String social_info) {
-            this.social_info = social_info;
-        }
-
-        public String getMember_info() {
-            return member_info;
-        }
-
-        public void setMember_info(String member_info) {
-            this.member_info = member_info;
-        }
-
-        public String getFamily_info() {
-            return family_info;
-        }
-
-        public void setFamily_info(String family_info) {
-            this.family_info = family_info;
-        }
-
-        public String getHealth_info() {
-            return health_info;
-        }
-
-        public void setHealth_info(String health_info) {
-            this.health_info = health_info;
-        }
-    }
-
-    public class ext1 {
-
-    }
-
-}
